+++ go/backend/archive/archive.go
@@ -13,19 +13,13 @@ package archive
 //go:generate mockgen -source archive.go -destination archive_mock.go -package archive
 
 import (
-	"io"
-
 	"github.com/Fantom-foundation/Carmen/go/common"
 	"github.com/Fantom-foundation/Carmen/go/common/amount"
 	"github.com/Fantom-foundation/Carmen/go/common/witness"
 )
 
-const (
-	// ErrWitnessProofNotSupported is returned when the archive does not support witness proofs.
-	ErrWitnessProofNotSupported = common.ConstError("witness proof not supported")
-	// ErrExportNotSupported is returned when the archive does not support export.
-	ErrExportNotSupported = common.ConstError("export not supported")
-)
+// ErrWitnessProofNotSupported is returned when the archive does not support witness proofs.
+const ErrWitnessProofNotSupported = common.ConstError("witness proof not supported")
 
 // An Archive retains a history of state mutations in a blockchain on a
 // block-level granularity. The history is recorded by adding per-block updates.
@@ -65,9 +59,6 @@ type Archive interface {
 
 	// CreateWitnessProof creates a witness proof for the given account and keys.
 	CreateWitnessProof(block uint64, address common.Address, keys ...common.Key) (witness.Proof, error)
-
-	// Export exports a LiveDB for given block into out.
-	Export(block uint64, out io.Writer) (common.Hash, error)
 
 	// MemoryFootprintProvider provides the size of the store in memory in bytes.
 	common.MemoryFootprintProvider
