+++ go/database/mpt/archive_trie.go
@@ -12,7 +12,6 @@ package mpt
 
 import (
 	"bufio"
-	"context"
 	"errors"
 	"fmt"
 	"io"
@@ -296,10 +295,6 @@ func (a *ArchiveTrie) GetMemoryFootprint() *common.MemoryFootprint {
 	mf.AddChild("roots", common.NewMemoryFootprint(uintptr(a.roots.length())*unsafe.Sizeof(NodeId(0))))
 	a.rootsMutex.Unlock()
 	return mf
-}
-
-func (a *ArchiveTrie) Export(block uint64, out io.Writer) (common.Hash, error) {
-	return ExportLive(context.Background(), NewExportableArchiveTrie(a, block), out)
 }
 
 func (a *ArchiveTrie) Check() error {
