+++ go/state/gostate/archive_state_test.go
@@ -11,7 +11,6 @@
 package gostate
 
 import (
-	"bytes"
 	"errors"
 	"fmt"
 	"testing"
@@ -103,15 +102,6 @@ func TestState_ArchiveState_FailingOperation_InvalidatesArchive(t *testing.T) {
 			},
 			func(stateArchive state.State) error {
 				_, err := stateArchive.CreateWitnessProof(common.Address{}, common.Key{})
-				return err
-			},
-		},
-		"liveDBGenesis": {
-			func(archive *archive.MockArchive, injectedErr error) {
-				archive.EXPECT().CreateLiveDBGenesis(gomock.Any(), gomock.Any()).Return(injectedErr)
-			},
-			func(stateArchive state.State) error {
-				_, err := stateArchive.Export(bytes.NewBuffer(nil))
 				return err
 			},
 		},
